+++ src/SPTQuestingBots.Client/Patches/AirdropLandPatch.cs
@@ -24,7 +24,7 @@ namespace SPTQuestingBots.Patches
         }
 
         [PatchPostfix]
-        protected static void PatchPostfix(AirdropSynchronizableObject ___airdropSynchronizableObject_0)
+        protected static void PatchPostfix(AirdropSynchronizableObject ___AirdropSynchronizableObject_0)
         {
             // Do not run this on Fika client machines
             if (!Helpers.RaidHelpers.IsHostRaid())
@@ -32,9 +32,9 @@ namespace SPTQuestingBots.Patches
                 return;
             }
 
-            AddNavMeshObstacle(___airdropSynchronizableObject_0);
+            AddNavMeshObstacle(___AirdropSynchronizableObject_0);
 
-            Vector3 airdropPosition = ___airdropSynchronizableObject_0.transform.position;
+            Vector3 airdropPosition = ___AirdropSynchronizableObject_0.transform.position;
             LoggingController.LogInfo(
                 "[AirdropLandPatch] Airdrop landed at ("
                     + airdropPosition.x.ToString("F0")
@@ -61,10 +61,10 @@ namespace SPTQuestingBots.Patches
             );
         }
 
-        private static void AddNavMeshObstacle(AirdropSynchronizableObject ___airdropSynchronizableObject_0)
+        private static void AddNavMeshObstacle(AirdropSynchronizableObject ___AirdropSynchronizableObject_0)
         {
-            NavMeshObstacle navMeshObstacle = ___airdropSynchronizableObject_0.gameObject.GetOrAddComponent<NavMeshObstacle>();
-            navMeshObstacle.size = ___airdropSynchronizableObject_0.CollisionCollider.bounds.size;
+            NavMeshObstacle navMeshObstacle = ___AirdropSynchronizableObject_0.gameObject.GetOrAddComponent<NavMeshObstacle>();
+            navMeshObstacle.size = ___AirdropSynchronizableObject_0.CollisionCollider.bounds.size;
             navMeshObstacle.carving = true;
         }
     }